Move `evaluations` app into `interviews` * Delete `evaluations` app * Create `interviewer` model as a many-to-many relationship between user and interview * Get rid of dublicate models * Add migrations * Register models in admin * Move `criterias/` endpoint/view
