--- python/apps/evaluations/admin.py
@@ -1,5 +1,0 @@
-from django.contrib import admin
-from apps.evaluations.models import Criteria, Evaluation, EvaluationCriteriaRelationship
-
-
-admin.site.register([Evaluation, Criteria, EvaluationCriteriaRelationship])
