--- python/apps/evaluations/urls.py
@@ -1,7 +1,0 @@
-from django.urls import path
-
-from apps.evaluations import views
-
-urlpatterns = [
-    path('criterias/', views.CriteriaCreateListView.as_view()),
-]
